# perf(ext/http): fast path for uncompressed bodies