Brings back the nice juicy numbers seen in https://github.com/denoland/deno/pull/14284

sample.txt is 5MB

`main`:
```
wrk -t8 -c20 -d10s http://localhost:8080/sample.txt
Running 10s test @ http://localhost:8080/sample.txt
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    38.43ms    4.80ms  83.32ms   91.27%
    Req/Sec    51.87      9.99    60.00     62.87%
  4192 requests in 10.10s, 20.53GB read
Requests/sec:    414.91
Transfer/sec:      2.03GB
```

This PR:
```
Running 10s test @ http://localhost:8080/sample.txt
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.58ms    2.53ms  44.05ms   88.56%
    Req/Sec   114.07     12.06   141.00     67.20%
  9179 requests in 10.10s, 44.87GB read
Requests/sec:    908.86
Transfer/sec:      4.44GB
```