@littledivy 2022-04-25
> What about a HttpResponseWriter::send() method that encapsulates logic shared across the two send functions?

Both logics are not exactly duplicated, one deals with an owned ZeroCopyBuf so it could avoid copy. `op_http_write` copies every time because of ownership outside the scope. Let's think of this in follow-ups.